refactor: Addressing Daniel's review feedback as part of PR (#3)
